A little bit of  clean-up
